src/cli/requirements.txt
typer==0.12.5
rich==13.9.4
httpx==0.27.2
pyyaml==6.0.2
